ref(seer): show conversation totals only and link query rows
Three changes to the conversation embeds:

The conversation block dropped its transcript. The embed renders inside
an agent conversation, so a nested transcript reads as part of the
surrounding answer; the LLM call, token, cost and error totals say what
the reader needs, and the link goes to the full detail view. Message
selection went with it, so the block is no longer interactive at all.

The conversationsQuery rows now link to their conversations, through the
same `getConversationHref` the conversation embed uses. They open in a
new tab so following one cannot replace the page holding the answer.

Layout now matches the other embeds: `conversationsQuery` is flat in
`components/` like every other query embed, `conversation/` keeps the
single-resource embed like `monitor/` and `alert/`, each embed has its
own colocated spec, and sibling imports are relative.
