+++ static/app/components/seer/markdown/embeds/index.ts
@@ -3,7 +3,7 @@ import {Alert} from './components/alert/alert';
 import {Autofix, AutofixRef} from './components/autofix';
 import {Chart} from './components/chart';
 import {Conversation} from './components/conversation/conversation';
-import {ConversationsQuery} from './components/conversation/conversationsQuery';
+import {ConversationsQuery} from './components/conversationsQuery';
 import {Dashboard} from './components/dashboard';
 import {Docs} from './components/docs';
 import {Dsn} from './components/dsn';
